Get the perfect itemDuration
Adjust the itemDuration to a little bit more than the time it takes to create and use 9 uraniumrods. This means that 1 uraniumrod crafter can supply 1 nuclear reactor and that forever. It will have about 2 rods left every time the reactor takes new ones so you could maybe use them accordingly but it is working as intended now.